requirements.txt
# Prisma AIRS Management SDK (TestPyPI)
# Install with: pip install --extra-index-url https://test.pypi.org/simple/ -r requirements.txt
pan-airs-api-mgmt-sdk==0.0.1a14
